+++ src/Components/Pages/Results/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Results(){
+    return <h1>Results</h1>
+}
