+++ src/Components/Pages/StartPage/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { Link} from "react-router-dom";
+
+export default function StartPage(){
+    return(
+        <Link to="/Typify">
+            <button>Start!</button>
+        </Link>
+    )
+}
