+++ src/Components/Pages/Typeify/Dummy/index.js
