+++ src/Components/Pages/Typeify/Input/index.js
