+++ src/Components/Pages/Typeify/Layout/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+export default function Layout(){
+    return(
+        <footer style={{
+            position:"absolute",
+              bottom:"3px" 
+            }}><a href="https://github.com/callumhemming/Typing-Speed-Test"><i>Check out the <u>source!</u> </i></a>
+            </footer>
+    )
+}
