+++ src/Components/Pages/Typeify/Timer/index.js
