+++ src/Components/Pages/Typeify/Typeify/App.css
@@ -1,0 +1,6 @@
+.container{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20%;
+}
